feat(优化): 首页、详情页代码优化
