+++ src/components/right/ProjectIntro.vue
@@ -1,18 +1,18 @@
 <template>
-  <div>
+  <div style="background-color: #fff">
     <a-row class="about_nsbbs_img">
-      <img src="@/assets/img/about-nsbbs.png" style="width: 100%;" />
+      <img src="@/assets/img/about-nsbbs.png" style="width: 100%" />
     </a-row>
     <a-row>
       <a-col :span="24" style="position: relative; top: -25px; text-align: center">
         <a-avatar :size="80" :src="require('@/assets/img/nan.jpg')" />
       </a-col>
       <a-col :span="24" style="text-align: center">
-        <div style="line-height: 28px; padding: 0 10px 10px 10px;">
+        <div style="line-height: 28px; padding: 0 10px 10px 10px">
           {{ $t('common.projectIntro') }}
         </div>
       </a-col>
-      <a-divider style="font-size: 12px;">{{ $t('common.commonCount') }}</a-divider>
+      <a-divider style="font-size: 12px">{{ $t('common.commonCount') }}</a-divider>
       <a-col :span="24" style="text-align: center">
         <a-col :span="6">
           <p>{{ $t('common.article') }}</p>
@@ -41,7 +41,6 @@
 import articleService from '@/service/articleService';
 
 export default {
-  props: {},
   data() {
     return {
       data: {},
@@ -62,7 +61,8 @@ export default {
 
     // 轮播图开关
     carouselSwitch(checked) {
-      this.$store.state.isCarousel = checked ? 1 : 0;
+      const isCarousel = checked ? 1 : 0;
+      this.$store.commit('setCarousel', isCarousel);
       window.localStorage.isCarousel = checked ? 1 : 0;
     },
   },
