+++ src/mixins/loadMoreWhenScroll.js
@@ -1,0 +1,7 @@
+import { scroll } from '@/utils';
+
+export const loadMoreWhenScroll = {
+  methods: {
+    scroll,
+  },
+};
